Falta comentar só uum bocado o resto, e talvez organizar melhora as classes
