+++ src/main/resources/scores/scores_level1.txt
@@ -1,3 +1,3 @@
 FFD;level1;28
 FFD;level1;28
-HHG;level1;30
+CCC;level1;28
