+++ src/main/resources/scores/scores_level2.txt
@@ -1,0 +1,3 @@
+SAZ;level2;41
+SAZ;level2;41
+AAZ;level2;42
